Address PR comments and modify menu header
